// src/reflect/Type.enum.ts
export enum Type {
  Array = "Array",
  AsyncFunction = "AsyncFunction",
  BigInt = "BigInt",
  Boolean = "Boolean",
  BooleanObject = "Boolean object",
  Function = "Function",
  FunctionObject = "Function object",
  Map = "Map",
  Null = "Null",
  Number = "Number",
  NumberObject = "Number object",
  Object = "Object",
  Promise = "Promise",
  Set = "Set",
  String = "String",
  StringObject = "String object",
  Symbol = "Symbol",
  Undefined = "undefined",
  WeakMap = "WeakMap",
  WeakSet = "WeakSet",
}